refactor: split credential discovery toolset into modular files
**Added:**

- Created `discovery.py` for low-hanging fruit credential discovery (LDAP, username=password, password spray, LAPS)
- Added `harvesting.py` for Active Directory credential harvesting (secretsdump, kerberoast, asrep roast, admin check)
- Introduced `cracking.py` for password hash cracking (hashcat and John the Ripper)
- Added `pilfering.py` for SMB share pilfering (spider, GPP passwords, SYSVOL search, NTDS extraction)

**Changed:**

- Updated `__init__.py` to re-export tool classes from new split modules and revised package structure documentation
- Updated all test patches in `test_sysvol_script_search.py` to reference new `pilfering` module instead of `_credential_discovery`
- Adjusted import paths and references throughout the codebase to accommodate new file structure

**Removed:**

- Removed monolithic `_credential_discovery.py` file containing all credential discovery tools in favor of modular design
